Pestañas Categorias, Ingredientes, Medidas, y Anotaciones terminadas

+++ RecetApp/src/dad/recetapp/ui/AnotacionesPanel.java
@@ -5,13 +5,19 @@ import java.util.List;
 
 import org.apache.pivot.beans.BXML;
 import org.apache.pivot.beans.Bindable;
+import org.apache.pivot.collections.ArrayList;
 import org.apache.pivot.collections.Map;
 import org.apache.pivot.collections.Sequence;
 import org.apache.pivot.util.Resources;
 import org.apache.pivot.wtk.Button;
 import org.apache.pivot.wtk.ButtonPressListener;
+import org.apache.pivot.wtk.MessageType;
+import org.apache.pivot.wtk.Prompt;
+import org.apache.pivot.wtk.Sheet;
+import org.apache.pivot.wtk.SheetCloseListener;
 import org.apache.pivot.wtk.TablePane;
 import org.apache.pivot.wtk.TableView;
+import org.apache.pivot.wtk.TableViewRowListener;
 import org.apache.pivot.wtk.TextInput;
 
 import dad.recetapp.services.ServiceException;
@@ -57,6 +63,13 @@ public class AnotacionesPanel extends TablePane implements Bindable {
 				onEliminarAnotacionButtonActionPerformed();
 			}
 		});
+		tableView.getTableViewRowListeners().add(new TableViewRowListener.Adapter(){
+		@Override
+			public void rowUpdated(TableView tableView, int row) {
+				onRowUpdated();
+				super.rowUpdated(tableView, row);
+			}	
+		});
 	}
 
 	protected void onAnadirAnotacionButtonActionPerformed() {
@@ -71,41 +84,42 @@ public class AnotacionesPanel extends TablePane implements Bindable {
 		nombreText.setText("");
 	}
 
-	protected void onEliminarAnotacionButtonActionPerformed() {
+	protected void onEliminarAnotacionButtonActionPerformed() {	
+		StringBuffer mensaje = new StringBuffer();
+		mensaje.append("¿Desea eliminar las siguientes anotaciones?\n\n");
+		
 		Sequence<?> seleccionados = tableView.getSelectedRows();
 		for (int i = 0; i < seleccionados.getLength(); i++) {
-			try {
-				TipoAnotacionItem anotacionSeleccionada = (TipoAnotacionItem) seleccionados.get(i);
-				variables.remove(anotacionSeleccionada);
-				ServiceLocator.getTiposAnotacionesService().eliminarTipoAnotacion(anotacionSeleccionada.getId());
-			} catch (ServiceException e) {
-				e.printStackTrace();
-			}
+			TipoAnotacionItem anotacionSeleccionada = (TipoAnotacionItem) seleccionados.get(i);
+			mensaje.append(" - " + anotacionSeleccionada.getDescripcion() + "\n");
 		}
 		
-		
-//		StringBuffer mensaje = new StringBuffer();
-//		mensaje.append("¿Desea eliminar las siguientes anotaciones?\n\n");
-//		
-//		Sequence<?> seleccionados = tableView.getSelectedRows();
-//		for (int i = 0; i < seleccionados.getLength(); i++) {
-//			TipoAnotacionItem variableSeleccionada = (TipoAnotacionItem) seleccionados.get(i);
-//			mensaje.append(" - " + variableSeleccionada.getDescripcion() + "\n");
-//		}
-//		
-//		Prompt confirmar = new Prompt(MessageType.WARNING, mensaje.toString(), new ArrayList<String>("Sí", "No"));
-//		confirmar.open(this, new SheetCloseListener() {
-//			public void sheetClosed(Sheet sheet) {
-//				
-//				if (confirmar.getResult() && confirmar.getSelectedOption().equals("Sí")) {
-//					Sequence<?> seleccionados = tableView.getSelectedRows();
-//					for (int i = 0; i < seleccionados.getLength(); i++) {
-//						variables.remove((TipoAnotacionItem)seleccionados.get(i));
-//					}
-//				}
-//				
-//			}
-//		});
+		Prompt confirmar = new Prompt(MessageType.WARNING, mensaje.toString(), new ArrayList<String>("Sí", "No"));
+		confirmar.open(this.getWindow(), new SheetCloseListener() {
+			public void sheetClosed(Sheet sheet) {
+				
+				if (confirmar.getResult() && confirmar.getSelectedOption().equals("Sí")) {
+					Sequence<?> seleccionados = tableView.getSelectedRows();
+					for (int i = 0; i < seleccionados.getLength(); i++) {
+						try {
+							TipoAnotacionItem anotacionSeleccionada = (TipoAnotacionItem) seleccionados.get(i);
+							variables.remove(anotacionSeleccionada);
+							ServiceLocator.getTiposAnotacionesService().eliminarTipoAnotacion(anotacionSeleccionada.getId());
+						} catch (ServiceException e) {
+							e.printStackTrace();
+						}
+					}	
+				}			
+			}
+		});
 	}
 
+	protected void onRowUpdated() {
+		TipoAnotacionItem seleccionado = (TipoAnotacionItem) tableView.getSelectedRow();
+		try {
+			ServiceLocator.getTiposAnotacionesService().modificarTipoAnotacion(seleccionado);
+		} catch (ServiceException e) {
+			e.printStackTrace();
+		}
+	}
 }
