+++ RecetApp/src/dad/recetapp/ui/IngredientesPanel.java
@@ -5,17 +5,24 @@ import java.util.List;
 
 import org.apache.pivot.beans.BXML;
 import org.apache.pivot.beans.Bindable;
+import org.apache.pivot.collections.ArrayList;
 import org.apache.pivot.collections.Map;
 import org.apache.pivot.collections.Sequence;
 import org.apache.pivot.util.Resources;
 import org.apache.pivot.wtk.Button;
 import org.apache.pivot.wtk.ButtonPressListener;
+import org.apache.pivot.wtk.MessageType;
+import org.apache.pivot.wtk.Prompt;
+import org.apache.pivot.wtk.Sheet;
+import org.apache.pivot.wtk.SheetCloseListener;
 import org.apache.pivot.wtk.TablePane;
 import org.apache.pivot.wtk.TableView;
+import org.apache.pivot.wtk.TableViewRowListener;
 import org.apache.pivot.wtk.TextInput;
 
 import dad.recetapp.services.ServiceException;
 import dad.recetapp.services.ServiceLocator;
+import dad.recetapp.services.items.TipoAnotacionItem;
 import dad.recetapp.services.items.TipoIngredienteItem;
 
 public class IngredientesPanel extends TablePane implements Bindable {
@@ -45,6 +52,7 @@ public class IngredientesPanel extends TablePane implements Bindable {
 			variables.add(l);
 		}
 		tableView.setTableData(variables);
+		
 		anadirIngredienteButton.getButtonPressListeners().add(new ButtonPressListener() {
 			@Override
 			public void buttonPressed(Button button) {
@@ -58,6 +66,14 @@ public class IngredientesPanel extends TablePane implements Bindable {
 				onEliminarIngredienteButtonActionPerformed();
 			}
 		});
+		
+		tableView.getTableViewRowListeners().add(new TableViewRowListener.Adapter(){
+			@Override
+				public void rowUpdated(TableView tableView, int row) {
+					onRowUpdated();
+					super.rowUpdated(tableView, row);
+				}	
+			});
 	}
 
 
@@ -76,15 +92,41 @@ public class IngredientesPanel extends TablePane implements Bindable {
 
 
 	protected void onEliminarIngredienteButtonActionPerformed() {
+		StringBuffer mensaje = new StringBuffer();
+		mensaje.append("¿Desea eliminar los siguientes tipos de ingrediente?\n\n");
+		
 		Sequence<?> seleccionados = tableView.getSelectedRows();
 		for (int i = 0; i < seleccionados.getLength(); i++) {
-			TipoIngredienteItem tipoIngredienteSeleccionado = (TipoIngredienteItem) seleccionados.get(i);
-			variables.remove(tipoIngredienteSeleccionado);
-			try {
-				ServiceLocator.getTiposIngredientesService().eliminarTipoIngrediente(tipoIngredienteSeleccionado.getId());
-			} catch (ServiceException e) {
-				e.printStackTrace();
+			TipoAnotacionItem tipoIngredienteSeleccionado = (TipoAnotacionItem) seleccionados.get(i);
+			mensaje.append(" - " + tipoIngredienteSeleccionado.getDescripcion() + "\n");
+		}
+		
+		Prompt confirmar = new Prompt(MessageType.WARNING, mensaje.toString(), new ArrayList<String>("Sí", "No"));
+		confirmar.open(this.getWindow(), new SheetCloseListener() {
+			public void sheetClosed(Sheet sheet) {
+				
+				if (confirmar.getResult() && confirmar.getSelectedOption().equals("Sí")) {
+					Sequence<?> seleccionados = tableView.getSelectedRows();
+					for (int i = 0; i < seleccionados.getLength(); i++) {
+						TipoIngredienteItem tipoIngredienteSeleccionado = (TipoIngredienteItem) seleccionados.get(i);
+						variables.remove(tipoIngredienteSeleccionado);
+						try {
+							ServiceLocator.getTiposIngredientesService().eliminarTipoIngrediente(tipoIngredienteSeleccionado.getId());
+						} catch (ServiceException e) {
+							e.printStackTrace();
+						}
+					}
+				}			
 			}
+		});
+	}
+	
+	protected void onRowUpdated() {
+		TipoIngredienteItem seleccionado = (TipoIngredienteItem) tableView.getSelectedRow();
+		try {
+			ServiceLocator.getTiposIngredientesService().modificarTipoIngrediente(seleccionado);
+		} catch (ServiceException e) {
+			e.printStackTrace();
 		}
 	}
 }
